--- components/brands/Brand/Brand.type.ts
@@ -1,5 +1,0 @@
-import { ContItems } from 'apis/models/Brands.type';
-
-export default interface BrandProps {
-  item: ContItems;
-}
