+++ components/category/Category.type.ts
@@ -1,0 +1,5 @@
+import { ContItems } from 'apis/models/Categories.type';
+
+export default interface CategoryProps {
+  item: ContItems;
+}
