--- pages/brands/Brands.type.ts
@@ -1,6 +1,0 @@
-import { ConCategory2s } from 'apis/models/Brands.type';
-
-export default interface BrandsProps {
-  name: string;
-  data: ConCategory2s[];
-}
