+++ pages/brands/[category2Id].tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { AppLayout } from 'components/common';
+
+const Brands = () => {
+  return (
+    <AppLayout title="test" backPath="/">
+      "test"
+    </AppLayout>
+  );
+};
+
+
+export default Brands;
+
