+++ pages/categories/Categories.styled.ts
