+++ pages/categories/Categories.type.ts
@@ -1,0 +1,6 @@
+import { ConCategory2s } from 'apis/models/Categories.type';
+
+export default interface CategoriesProps {
+  name: string;
+  data: ConCategory2s[];
+}
