--- pages/categories/[category2Id].tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { AppLayout } from 'components/common';
-const Categories = () => {
-  return (
-    <AppLayout title="카테고리" backPath="/">
-      Categories
-    </AppLayout>
-  );
-};
-
-export default Categories;
